requirements.txt
aiohttp==3.8.4
aiohttp-session==2.12.0
aiosmtpd==1.4.4
aiosqlite==0.18.0
python-dotenv==1.0.0